requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.35.3
pillow==10.3.0
python-dotenv==1.0.1
pydantic==2.7.4